all functions are gone from main, the folder for Blackjack needs to be changed to RPS. RPS is game2. I did not know how to create new folders in games, so I reused existing architecture to make it work
