+++ src/include/roulette.h
@@ -1,0 +1,10 @@
+#ifndef ROULETTE_H
+#define ROULETTE_H
+#include <SFML/Graphics.hpp>
+#include "button.h"
+#include <iostream>
+
+void game1();
+
+
+#endif
