+++ src/include/rps.h
@@ -1,0 +1,11 @@
+#ifndef RPS_H
+#define RPS_H
+#include <SFML/Graphics.hpp>
+#include "button.h"
+#include <iostream>
+
+void game2();
+int botChoice();
+int RPS(int user, int computer);
+
+#endif
